making the changes requested
